requirements.txt
streamlit
pandas 
json
datetime
plotly.express 